chore: isolate unit tests for custom types by numpyro
